+++ Cartfile
@@ -1,6 +1,6 @@
 github "wireapp/wire-ios-ziphy" ~> 14.0
 github "wireapp/Cartography" "4.0.0-xcode_11_4_1"
-github "wireapp/wire-ios-sync-engine" ~> 342.0
+github "wireapp/wire-ios-sync-engine" ~> "feature/ZIOS-13827_DeepLinkFix"
 github "wireapp/wire-ios-share-engine" ~> 193.0
 github "wireapp/FormatterKit" "1.8.1-swift3.0.2"
 github "wireapp/wire-ios-canvas" ~> 11.0
